Create One Stick Drive
I think I figured out a way to do the one stick control, but it's not pretty.  The best way I could solve the issue with going above 100% and below -100% in quadrants 1 and 3, respectively, was to change them to the maximum velocity if it reached that.  The idea behind the program is to allow the user as much control over the speed/direction of the robot as possible.  If it turns to the right, the left motor can still move at whatever velocity the joystick is being pushed up, on Axis 3, and the right motor must slow down based on how far the stick is pushed in that direction.  The same principle applies to turning right in reverse.